Add image support

+++ src/main/java/com/rebalance/service/ImageService.java
@@ -7,7 +7,9 @@ import com.rebalance.entity.Image;
 import com.rebalance.exception.RebalanceErrorType;
 import com.rebalance.exception.RebalanceException;
 import com.rebalance.repository.ImageRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import jakarta.annotation.PostConstruct;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
@@ -16,12 +18,18 @@ import java.awt.image.BufferedImage;
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
 import java.util.Base64;
 
+import static java.util.Objects.nonNull;
+import static org.springframework.util.Assert.notNull;
+
+@Slf4j
 @Service
+@RequiredArgsConstructor
 public class ImageService {
 
+    public static final String PNG = ".png";
+    public static final String ICON_JPG = "_icon.jpg";
     @Value("${cloud.storage.account.name}")
     private String storageAccountName;
 
@@ -37,64 +45,60 @@ public class ImageService {
     private String imageNamePrefix;
 
     private final ImageRepository imageRepository;
+    private BlobServiceClient connection;
 
-    @Autowired
-    public ImageService(ImageRepository imageRepository) {
-        this.imageRepository = imageRepository;
+    @PostConstruct
+    @SuppressWarnings("unused")
+    public void init() {
+        connection = connectToCloudStorage();
     }
 
-    public void saveImage(String base64Image, Long globalId) {
-        String uri = saveImageAndIconToCloud(base64Image, globalId);
+    public void saveImage(Long globalId, byte[] image) {
+        String uri = saveImageAndIconToCloud(image, globalId);
         Image toRepository = new Image(globalId, uri);
         imageRepository.save(toRepository);
     }
 
-    public String getImageByGlobalId(Long id) {
-        throwExceptionIfImageNotExists(id);
-        byte[] imageBytes = loadImageBytes(id, false);
-        return Base64.getEncoder().encodeToString(imageBytes);
-    }
-
-    public String getImageIconByGlobalId(Long id) {
-        throwExceptionIfIconNotExists(id);
-        byte[] imageBytes = loadImageBytes(id, true);
-        return Base64.getEncoder().encodeToString(imageBytes);
-    }
-
     public void deleteImageByGlobalId(Long id) {
-        throwExceptionIfImageNotExists(id);
-
-        BlobServiceClient connection = connectToCloudStorage();
-
-        connection.getBlobContainerClient(storageContainerImages).getBlobClient(id + imageNamePrefix + ".png").delete();
-        connection.getBlobContainerClient(storageContainerThumbnails).getBlobClient(id + imageNamePrefix + "_icon.jpg").delete();
+        if (!imageExistsByGlobalId(id)) {
+            return;
+        }
+        connection.getBlobContainerClient(storageContainerImages).getBlobClient(id + imageNamePrefix + PNG).delete();
+        connection.getBlobContainerClient(storageContainerThumbnails).getBlobClient(id + imageNamePrefix + ICON_JPG).delete();
 
         imageRepository.deleteById(id);
     }
 
-    public void throwExceptionIfImageNotExists(Long id) {
-        imageRepository.findById(id).orElseThrow(() -> new RebalanceException(RebalanceErrorType.RB_301));
+    public void throwExceptionIfMediaNotExists(Long globalId) {
+        if (imageRepository.findById(globalId).isEmpty()) {
+            throw new RebalanceException(RebalanceErrorType.RB_301);
+        }
     }
 
-    public void throwExceptionIfIconNotExists(Long id) {
-        imageRepository.findById(id).orElseThrow(() -> new RebalanceException(RebalanceErrorType.RB_302));
+    public boolean imageExistsByGlobalId(Long globalId) {
+        return imageRepository.findById(globalId).isPresent();
     }
 
-    public void updateImage(Long id, String base64Image) {
-        BlobServiceClient connection = connectToCloudStorage();
-
-        connection.getBlobContainerClient(storageContainerImages).getBlobClient(id + imageNamePrefix + ".png").delete();
-        connection.getBlobContainerClient(storageContainerThumbnails).getBlobClient(id + imageNamePrefix + "_icon.jpg").delete();
-
-        saveImageAndIconToCloud(base64Image, id);
+    public void updateImage(Long globalId, byte[] image) {
+        deleteImageByGlobalId(globalId);
+        saveImage(globalId, image);
     }
 
-    private byte[] loadImageBytes(Long globalId, boolean getIcon) {
-        BlobServiceClient connection = connectToCloudStorage();
-        String container = getIcon ? storageContainerThumbnails : storageContainerImages;
-        String blobName = getIcon ? globalId + imageNamePrefix + "_icon.jpg" : globalId + imageNamePrefix + ".png";
-        return connection.getBlobContainerClient(container).getBlobClient(blobName).downloadContent().toBytes();
+    public String loadImageBase64(Long globalId, boolean getIcon) {
+        if (imageRepository.findById(globalId).isEmpty()) {
+            return null;
+        }
+        try {
+            String container = getIcon ? storageContainerThumbnails : storageContainerImages;
+            String blobName = getIcon ? globalId + imageNamePrefix + ICON_JPG : globalId + imageNamePrefix + PNG;
+            byte[] imageBytes = connection.getBlobContainerClient(container).getBlobClient(blobName).downloadContent().toBytes();
+            return nonNull(imageBytes) ? Base64.getEncoder().encodeToString(imageBytes) : null;
+        } catch (Exception e) {
+            log.error(e.toString());
+            return null;
+        }
     }
+
 
     private BlobServiceClient connectToCloudStorage() {
         return new BlobServiceClientBuilder()
@@ -103,22 +107,18 @@ public class ImageService {
                 .buildClient();
     }
 
-    private String saveImageAndIconToCloud(String base64Image, Long globalId) {
+    private String saveImageAndIconToCloud(byte[] imageBytes, Long globalId) {
         try {
-            BlobServiceClient connection = connectToCloudStorage();
+            String blobName = globalId + imageNamePrefix + PNG;
+            connection.getBlobContainerClient(storageContainerImages).getBlobClient(blobName).upload(BinaryData.fromBytes(imageBytes));
 
-            byte[] decodedImage = Base64.getDecoder().decode(base64Image.getBytes(StandardCharsets.UTF_8));
-
-            String blobName = globalId + imageNamePrefix + ".png";
-            connection.getBlobContainerClient(storageContainerImages).getBlobClient(blobName).upload(BinaryData.fromBytes(Base64.getDecoder().decode(base64Image.getBytes(StandardCharsets.UTF_8))));
-
-            BufferedImage img = ImageIO.read(new ByteArrayInputStream(decodedImage));
+            BufferedImage img = ImageIO.read(new ByteArrayInputStream(imageBytes));
+            notNull(img, "Invalid image data");
             java.awt.Image scaledImage = img.getScaledInstance(100, 100, java.awt.Image.SCALE_REPLICATE);
-
             BufferedImage bufferedImage = new BufferedImage(scaledImage.getWidth(null), scaledImage.getHeight(null), BufferedImage.TYPE_INT_RGB);
             bufferedImage.getGraphics().drawImage(scaledImage, 0, 0, null);
 
-            String iconBlobName = globalId + imageNamePrefix + "_icon.jpg";
+            String iconBlobName = globalId + imageNamePrefix + ICON_JPG;
 
             ByteArrayOutputStream b = new ByteArrayOutputStream();
             ImageIO.write(bufferedImage, "jpg", b);
@@ -126,7 +126,11 @@ public class ImageService {
             connection.getBlobContainerClient(storageContainerThumbnails).getBlobClient(iconBlobName).upload(BinaryData.fromBytes(b.toByteArray()));
             return String.format("https://%s.blob.core.windows.net/%s/%s", storageAccountName, storageContainerImages, blobName);
         } catch (IOException e) {
-            throw new RuntimeException("Error saving image", e);
+            log.error("Error saving image [globalId={}]", globalId);
+            throw new RebalanceException(RebalanceErrorType.RB_302);
+        } catch (IllegalArgumentException e) {
+            log.error("Invalid image data [globalId={}]", globalId);
+            throw new RebalanceException(RebalanceErrorType.RB_303);
         }
     }
